+++ js/views.js
@@ -2,7 +2,7 @@
  * @name views
  * @namespace Contains all the view modules.
  */
-define(["views/dashboard", "views/deleteLeague", "views/draft", "views/draftSummary", "views/exportRosters", "views/freeAgents", "views/gameLog", "views/hallOfFame", "views/editTeamInfo", "views/history", "views/inbox", "views/leaders", "views/leagueDashboard", "views/leagueFinances", "views/manual", "views/message", "views/negotiation", "views/negotiationList", "views/newLeague", "views/newTeam", "views/player", "views/playerRatingDists", "views/playerRatings", "views/playerShotLocations", "views/playerStatDists", "views/playerStats", "views/playoffs", "views/roster", "views/schedule", "views/standings", "views/teamFinances", "views/teamHistory", "views/teamShotLocations", "views/teamStatDists", "views/teamStats", "views/trade"], function (dashboard, deleteLeague, draft, draftSummary, exportRosters, freeAgents, gameLog, hallOfFame, editTeamInfo, history, inbox, leaders, leagueDashboard, leagueFinances, manual, message, negotiation, negotiationList, newLeague, newTeam, player, playerRatingDists, playerRatings, playerShotLocations, playerStatDists, playerStats, playoffs, roster, schedule, standings, teamFinances, teamHistory, teamShotLocations, teamStatDists, teamStats, trade) {
+define(["views/dashboard", "views/deleteLeague", "views/draft", "views/draftSummary", "views/exportRosters", "views/freeAgents", "views/gameLog", "views/hallOfFame", "views/editTeamInfo", "views/history", "views/inbox", "views/leaders", "views/leagueDashboard", "views/leagueFinances", "views/manual", "views/message", "views/negotiation", "views/negotiationList", "views/newLeague", "views/newTeam", "views/player", "views/playerRatingDists", "views/playerRatings", "views/playerShotLocations", "views/playerStatDists", "views/playerStats", "views/playoffs", "views/roster", "views/schedule", "views/standings", "views/teamFinances", "views/teamHistory", "views/teamShotLocations", "views/teamStatDists", "views/teamStats", "views/trade", "views/tradingBlock"], function (dashboard, deleteLeague, draft, draftSummary, exportRosters, freeAgents, gameLog, hallOfFame, editTeamInfo, history, inbox, leaders, leagueDashboard, leagueFinances, manual, message, negotiation, negotiationList, newLeague, newTeam, player, playerRatingDists, playerRatings, playerShotLocations, playerStatDists, playerStats, playoffs, roster, schedule, standings, teamFinances, teamHistory, teamShotLocations, teamStatDists, teamStats, trade, tradingBlock) {
     "use strict";
 
     return {
@@ -42,6 +42,7 @@ define(["views/dashboard", "views/deleteLeague", "views/draft", "views/draftSumm
         teamStatDists: teamStatDists,
         playerShotLocations: playerShotLocations,
         teamShotLocations: teamShotLocations,
-        exportRosters: exportRosters
+        exportRosters: exportRosters,
+        tradingBlock: tradingBlock
     };
 });